Fix KASAN

+++ inc/memlayout.h
@@ -217,7 +217,7 @@ struct PageInfo {
    * boot_alloc do not have valid reference count fields
    */
 
-  uint16_t pp_ref;
+  uint32_t pp_ref;
 };
 
 #endif /* !__ASSEMBLER__ */
